Add route to process images
This commit adds a route to process images, it also adds the pipeline to process image (media) and get text from it. The pipeline when the media is transformed to text is the same as the text one. The only difference is the creation of an additional ES entry that associates the media hash to a text hash.